Create balanced_nutrition_recommendation.ipynb
Model generated through using given nutrition dataset and ridge regression method. The idea was to use the available items in the area and create a balanced nutrition recommendation for farmers to feed the cow. The output generated tells the farmer the type of feed farmer needs to feed their bovine. For better results reinforcement learning and multiple ridge regression needs to be employed with larger dataset.